Feat: *add the lack test case and corresponding        relationships in cfi category       *further, adjust dependency relationships
