+++ cfi/call-fake-vtable-arg-num.cpp
