+++ cfi/call-fake-vtable-arg-type-modified.cpp
