+++ cfi/call-fake-vtable-arg-type.cpp
