+++ cfi/call-fake-vtable.cpp
