+++ cfi/call-wrong-vtable-arg-num.cpp
