+++ cfi/call-wrong-vtable-arg-type.cpp
