+++ cfi/call-wrong-vtable-child.cpp
