+++ cfi/call-wrong-vtable-func-num.cpp
